pyproject.toml
[tool.poetry]
name = "sabia"
version = "0.1.0"
description = ""
authors = ["Ana Paula Gomes <apgomes88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
